build: Add bazel wrapper
This wrapper is a simple shell script that handles downloading bazel.
This ensures that all developers are using the same supported version of
bazel, as well as keeping it in lockstep with CI. This also means
developers who aren't already using bazel don't have to figure out how
to install it on their system globally.

Signed-off-by: Keith Smiley <keithbsmiley@gmail.com>