swift: throw when setting restricted headers
Since we expose headers in the `Request` type, it's technically possible for consumers to set restricted headers that we will be setting internally within the library (any header starting with `:` in this case - `:method` for example).

To prevent this, we will throw an error when any of these headers/trailers is set manually.

Context: https://envoyproxy.slack.com/archives/CKQ2LK23G/p1563826656004700

Signed-off-by: Michael Rebello <mrebello@lyft.com>