swift: expose URL components instead of URL type
As we bridge over the platform layers to talk to the C/C++ core, we'll need to be able to send headers such as `:method`/`:scheme`/`:authority`.

The `URL` type exposes optionals for these properties, which means we need to either handle them at initialization time or at the time when the request is executed by unwrapping them and throwing/failing.

In order to prevent throwing when deconstructing `URL` components, we are instead updating the interfaces to accept these 3 fields separately.

Signed-off-by: Michael Rebello <mrebello@lyft.com>