ios: update unary/streaming & add test
Per the discussion [here](https://github.com/lyft/envoy-mobile/pull/312#discussion_r315820916), updating the unary and streaming interfaces.

Also moving the non-streaming convenience function into an extension on the protocol type so that it's available to all consumers.

Lastly, added a test to validate the default behavior of this extension function.

Signed-off-by: Michael Rebello <me@michaelrebello.com>