ios/core: data passthrough (#391)
Description: this PR adds functionality to send data to HTTP streams and receive data from callbacks. It also wires up this functionality on the ios platform code.
Risk Level: med - adds core functionality.
Testing: tested locally with the swift app and an echo endpoint. Data was seen on both the send and the received direction. Unit tests for the core code will be done in a subsequent PR.

Fixes #284

Co-authored-by: Mike Schore mike.schore@gmail.com
Signed-off-by: Jose Nino jnino@lyft.com