objective-c: guard against nil callbacks
Right now, if a consumer fails to set any closure in the `EnvoyHTTPCallbacks`, a crash will occur when the `EnvoyHTTPStreamImpl` attempts to call the closure.

It's reasonable to assume that a consumer might not care about each available callback (i.e., never setting `onCanceled`). We can guard against this by validating each closure is actually set before trying to call it.

Signed-off-by: Michael Rebello <me@michaelrebello.com>