swift: make RequestMethod.stringValue public
This can be useful to external consumers since the names of the enum cases are not convertible to strings due to the `@objc` attribute.

For example, when printing `print(RequestMethod.get)`, this will log `RequestMethod` right now since Objective-C int-backed enums don't store the names of Swift cases. Adding this read-only string representation allows consumers to utilize it.

Signed-off-by: Michael Rebello <me@michaelrebello.com>