kotlin: clear error closure after compression error
In the case where the client receives an unsupported compressed message, the `errorClosure` should be cleared out with the other handlers after being called once. This will prevent it from being called again if multiple compressed messages are received, which is the intended behavior as `onError` should be considered a terminal state.

Signed-off-by: Michael Rebello <me@michaelrebello.com>