build: move rules_swift repo rule earlier (#638)
Description:

We had reports of debugging issues due to paths that don't exist on developer machines. Previously we enabled the `swift.cacheable_modules` feature to prevent this (see #549). Apparently `rules_foreign_cc_dependencies()` is loading `rules_swift`, and must use a sha that is prior to the introduction of the `cacheable_modules` feature.

This change moves the `rules_swift` repository rule earlier in the `WORKSPACE` – above `rules_foreign_cc_dependencies()` and just below the `rules_appe` repo rule.

Risk Level:

Low

The rules_swift declaration was previously inert, it wasn't being used, so this change is effectively a bump to the rules_swift version being used/dependend on.

Testing:

Used `bazel aquery //library/swift/src:ios_framework_archive` and verified that `-no-serialize-debugging-options` was present, and that `-serialize-debugging-options` was **not included**.

Docs Changes: N/A

Release Notes: N/A