envoy: move ref back to stable (#648)
Signed-off-by: Jose Nino jnino@lyft.com
Description: envoyproxy/envoy#9618 broke the iOS build due to missing symbols. envoyproxy/envoy#9875 fixes. However, in order to expedite a clean master branch this PR moves the Envoy ref back to a stable place. Note that Android logging is reverted. Also note that CI for iOS was not testing for liveliness, which is how the breakage went through in the first place. This PR also fixes that.
Risk Level: low
Testing: CI

Fixes #646

Signed-off-by: Jose Nino <jnino@lyft.com>