platform: close streams explicitly with data or trailers
Replaces the `close(trailers: Map<String, List<String>>?)` function with two new functions (one for closing with data, and one for closing with trailers).

This removes the ambiguity of how a stream will be closed by making the consumer specify it explicitly.

The change is being done as a precursor for integrating L7 Swift/Kotlin filters, which will have interfaces in which trailers are not nullable.

Signed-off-by: Michael Rebello <me@michaelrebello.com>