Swift: Workaround Swift 5.1 swiftinterface bug
With Swift < 5.2 there's an issue where if you implicitly inheret
initializer from `NSObject` the generated swiftinterface doesn't include
the `override` keyword for subclasses and fails when trying to import
the framework. This makes that explicit which adds the keyword. Once we
upgrade to build with Swift 5.2 we can revert this.

Signed-off-by: Keith Smiley <keithbsmiley@gmail.com>