swift: update stream client to use factory closures for filters
Updates the stream client builder to accept a closure returning a `Filter` rather than requiring `Filter` to have an `init()` in its protocol conformance. This allows consumers to provide more flexible implementations of "factories" (closures) that can be used with custom initializers or alongside additional logic.

Note: Generics are required here since `Filter.name` is needed.

As part of this change, I also fixed some outdated documentation of this function.

Signed-off-by: Michael Rebello <me@michaelrebello.com>