http dispatcher: serialize all dispatch and state management (#1056)
Description: Previously cancel had the contract that by the time a call to `cancel` returned, no further dispatch would occur for a stream. This was guarded by some atomic state on the stream. While this was actually one of the more tested areas of the codebase, we're unwinding that for now, and dispatching cancellation the same as any other stream interaction. The hope is this will simplify reasoning about the code to continue investigation of recent crashes. Some tests have been removed that specifically tested concurrency of cancellation, since ordering tests now cover this.
Risk Level: Moderate
Testing: Local and CI

Signed-off-by: Mike Schore <mike.schore@gmail.com>