core/platform: add platform-level postinit callback wiring (#1087)
This PR adds a new set of `onEngineRunning` callbacks to the platform (Swift/Kotlin) interfaces.

The intent behind these callbacks is to allow platform consumers to observe when Envoy Mobile has finished setting itself up internally. When `build()` is called from an `EnvoyEngineBuilder`, internal library setup is not complete when the allocation returns. Instead, Envoy actually does additional setup before being called back with a "post init" step from the core Envoy layer indicating that it's ready to send requests. In practice, this doesn't have any direct effect on consumers since any requests sent into the library during this timeframe are queued for sending automatically once setup is complete.

Based on my localized runs of the library, I noticed 800-1000ms between the start of allocation and the time at which Envoy Mobile has completed its setup. There is room for improvement here which will be looked to in the future.

Changes:
- Add the new interface to platform layers
- Implement the core wiring for this callback and pipe that all the way through JNI / Objective-C up to Kotlin/Swift
- Update sample apps to utilize this functionality
- Add core unit test to validate the core callback completes

Signed-off-by: Michael Rebello <me@michaelrebello.com>
Co-authored-by: Alan Chiu <achiu@lyft.com>