config: rename decompressor name from "basic" to "gzip"
This filter is configured exclusively with gzip decompression. If we were to add more in the future (i.e., brotli), this would be a good way to differentiate.

This change is especially useful since it'll transform `x-envoy-decompressor-basic-uncompressed-bytes` to use "gzip" instead of "basic" in the trailers it adds to responses.

Signed-off-by: Michael Rebello <me@michaelrebello.com>