python: implement bridge utilities (#1222)
Description: This implements all of the util functions used to convert between envoy-mobile types and cc-lib types. Note that right now it copies _all_ data at the border between envoy-mobile and the cc-lib, which is inefficient. As noted in the TODO, this is going to be a good place for optimization.

Risk Level: Low
Testing: N/A, will follow up with testing in only C++, since this won't be exposed into the Python layer.
Docs Changes: N/A
Release Notes: N/A

Signed-off-by: Cerek Hillen <chillen@lyft.com>