python: expose terminate method (#1319)
Description: Based on some testing, if an `Engine` is not free'd before a Python process exits the it will crash on exit. This can happen while globally caching an instance of the `Engine`, for example. This PR exposes a `terminate` method that can be called on exit with [atexit](https://docs.python.org/3/library/atexit.html) for example.
Risk Level: Low
Testing: Pending

Signed-off-by: Cerek Hillen <chillen@lyft.com>