python: wheel building for envoy_engine (#1344)
Description: Bumps `rules_python` to version `0.2.0` and uses that new version to build a wheel for `envoy_engine.so`.  This is a test run for building an envoy_requests wheel down the road when it's upstreamed.
Risk Level: Low
Testing: `build //library/python:envoy_engine_whl`
Docs Changes: N/A
Release Notes: N/A

Signed-off-by: Cerek Hillen <chillen@lyft.com>
Signed-off-by: Charles Le Borgne <cleborgne@google.com>