bugfix: ensure PBFs pending destruction can't be resumed async (#1488)
Description: Previously, asynchronous filter resumption of a bridged platform filter was guarded in the PBF by a `weak_ptr` that ensured the object was still around. It turns out this is not sufficient, as filters are submitted for delayed destruction and the presence of the object isn't enough to ensure it's still associated with a functional stream. This fix adds an additional check to ensure we don't attempt to resume iteration on a PlatformBridgeFilter that is pending destruction.
Risk Level: Low
Testing: Unit and CI

Signed-off-by: Mike Schore <mike.schore@gmail.com>