Port CronetUrlRequestTest (#1538)
This PR is the starting point to port Cronet tests - it includes the main Cronet test: CronetUrlRequestTest. However, most the the code is for the test infrastructure itself: the CronetTestRule, NativeTestServer, MockUrlRequestJobFactory and some others. At this point, most of them are "fake" implementations. For example, the NativeTestServer has nothing "native" - it is just a wrapper around MockWebServer, which was coded to behave as much as possible like the original C++ Cronet Test Server.

With this PR, rhe CronetUrlRequestTest only tests the Java implementation (65 tests). Testing with Envoy Mobile incurs 16 test failures, so it is disabled for the moment. This will be mitigated with other PRs.

Description:  Port enough Cronet code to have test CronetUrlRequestTest running
Risk Level: None
Testing: this PR objective is to port a test
Docs Changes: N/A
Release Notes: N/A
Signed-off-by: Charles Le Borgne <cleborgne@google.com>