CI: Add linux config for Engflow's remote execution (#1709)
Description: Add linux config for Engflow's remote execution. No semantic changes. This PR is a no-op.

We add two things to the envoy-mobile build:
- A Linux-specific bazelrc configuration that sets all the necessary flags to run on a Linux Engflow worker
- Instruct envoy to use the latest envoyproxy/envoy-build-ubuntu docker container for the linux builds through the WORKSPACE rules

Notice that the envoy image is only necessary for the Linux builds since the MacOS builds are set somehow manually and don't use containers. Also notice that Engflow builds have a notion of "Pool" which for now only apply to Linux workers since the first workers in the Envoy cluster were Macs and they didn't have a Pool associated to it (e.g. no pool means Mac at the moment). Once the MacOS workers belong to a Pool we'll update the MacOS config.

Risk Level: Low
Testing: No-op
Docs Changes: N/A
Release Notes: N/A

Signed-off-by: Luis Fernando Pino Duque <luis@engflow.com>