emit events based on ENVOY_LOG_EVENT
Wires up the EM custom logger such that we intercept logs with stable names emitted by
ENVOY_LOG_EVENT, translating them into an event passed to the event tracker.

To maintain the original behavior when a logger is not installed, we end up having to partially
reimplement the Stderr logger.

Signed-off-by: Snow Pettersen <snowp@lyft.com>