Remove Kotlin dependency from Cronvoy (#1808)
It turns out that removing Kotlin from Envoy Mobile is not really needed at this point. In reality, the Kotlin classes are a layer on top of the Java classes, and Cronvoy is a layer too: one level of indirection for the callbacks is enough. So instead Cronvoy uses the java classes directly. The main drawback is the CronvoyConfiguration - there is some duplication. 

Description: Have Cronvoy to only use the Envoy Mobile Java classes for its implementation  
Risk Level: None
Testing: CI
Docs Changes: N/A
Release Notes: N/A
Fixes #1658
Signed-off-by: Charles Le Borgne <cleborgne@google.com>