Allow ByteBuffer.position() to be used for the data length. (#1906)
Description: Today, when sending the request body, the provided ByteBuffer is read entirely - its internal pointers such as the `position` and the `limit` are ignored. For the Cronvoy implementation, this is unfortunate. The lifecycle of the ByteBuffers used to send data are fully controlled by Cronet. Therefore, Cronet creates a ByteBuffer without knowing how many bytes the user will put in it. This can be made to work, but before sending the data, a new ByteBuffer needs to be created with the actual size that was put in the original ByteBuffer.

This PR attempts to solve this. By default, and this won't change, `ByteBuffer.capacity()` is used to determine the length when invoking `EnvoyHTTPStream.sendData()`. When configuring a new Stream, the user can now specify what should be used for the "length".  `StreamPrototype.setUseByteBufferPosition` is the new method allowing to rather use the `ByteBuffer.position()` as the length. Then, for Cronet, no need to copy the ByteBuffer with this strategy.

Technical note: the general expectation with a ByteBuffer is to rely on its `limit` and its `position` when accessing the data. Relying solely on its capacity is somewhat unorthodox. The proposed implementation with this PR is a bit more in line with the ByteBuffer intended usage, but is not "by the book" either: Normally, when reading, the ByteBuffer must be "flipped", and then each time bytes are consumed, the position must be increased accordingly. This implementation leaves the ByteBuffer untouched when sending. Trying to perfectly mimic this behavior might be tricky: the C++ layer would need to update the ByteBuffer's position once some data has been consumed. For Cronet, this lack of conformity is not concerning: the ByteBuffer lifecycle is fully controlled by Cronet, not by the user - by contract, the user only puts the data it wants to send in the Cronet provided ByteBuffer, strictly when asks to do so (through callback), nothing else.

Risk Level: Small: the business logic is the same when the feature is not used, which is the default.
Testing: CI, Unit Test added
Docs Changes: N/A
Release Notes: N/A

Signed-off-by: Charles Le Borgne <cleborgne@google.com>