[EnvoyBridgeUtility] Fix NSString to envoy_data conversion (#1958)
Description: Fix NSString to envoy_data conversion for non-ascii strings such as '台灣大哥大'. -length is the number of UTF-16 code units in the receiver whereas -lengthOfBytesUsingEncoding: is the number of UTF8 bytes. Using the latter is necessary since we're using this value to determine how many bytes to copy from the string.
Risk Level: Low.
Testing: Added an Objective-C unit test, and enabled running on CI.

Signed-off-by: JP Simard <jp@jpsim.com>