size: optimizations for dynamic library (#2049)
* size: compile releases optimizing for code size

Size delta: -0.66 Mi

* size: set default visibility to hidden

The .dynstr section of the binary as well as the symbol table is pretty big
for a shared library with almost no (intended to be) exported symbols. This
change makes all symbols hidden by default. We should test and add to visible
if some are missing.

Size delta: -5.66Mi

* size: override pgv validations with NOP functions

This change updates pgv to a version that supports a "cc nop" language. With
that we can make the validation functions a NOP that hopefuly gets inlined.

This is the PGV change for reference: envoyproxy/protoc-gen-validate#533

Size delta: -3.7 Mi

Signed-off-by: Martin Conte Mac Donell <reflejo@gmail.com>