[iOS/macOS] Remove `EngineBuilder.enableNetworkPathMonitor(_:)`
We ran some experiments in the Lyft apps and found no regressions with this
enabled, so we believe it's safe to remove the legacy `SCNetworkReachability`
implementation and always use the `NWPathMonitor` code path.

Also enable path monitor reachability state observation on macOS while I'm in
there.

It was disabled in https://github.com/envoyproxy/envoy-mobile/pull/2052 because
we didn't have a minimum macOS version set but this API is available as of
10.14.

Signed-off-by: JP Simard <jp@jpsim.com>