engine: return new handles from `EngineHandle::initEngine`
Until now, `EngineHandle` has always had a single value of `1`. In other
words, the Envoy engine has always been a singleton with no way to run
multiple independent engines.

In order to support multiple concurrent engines, we first need to make
`EngineHandle::initEngine` create a new engine and rather than have
functions that use the engine do a singleton lookup, we can use the
`envoy_engine_t` as an opaque handle type to get back the
`Envoy::Engine` instance.

Several places were using a hardcoded value of `1`, so we need to plumb
through the engine handles.

With this change although you can now create multiple engine instances,
it is not safe to do so because there is still some static state that
will need to be untangled. Also note that like the singleton we have had
until now, new engine instances that are created will leak, as there is
no way to fully release the engine through the public API. This is
intentional in order to reduce the risk associated with this change.

Co-authored-by: Mike Schore <mike.schore@gmail.com>
Co-authored-by: JP Simard <jp@jpsim.com>

Signed-off-by: JP Simard <jp@jpsim.com>