h3/quic: add experimental option to the Android/JVM EngineBuilder (#2163)
Description: Adds *experimental* support for negotiating H3/QUIC connections. To enable this, `enableHttp3` must be set to true on the EngineBuilder AND the header "x-envoy-mobile-upstream-protocol" must be set to "http3" on the request. If the upstream does not support H3, the standard ALPN over TCP configuration will still be used by default.
Risk Level: Moderate
Testing: Updated Unit Coverage

Co-authored-by: Alyssa Wilk <alyssar@chromium.org>
Signed-off-by: Mike Schore <mike.schore@gmail.com>