extensions: register `TestAccessor` filter
We use this in Lyft's integration with Envoy Mobile. It should be
harmless to add even though we don't currently need it, and it will be
helpful to keep the Envoy Mobile configuration close to the Lyft one
considering that we're running Envoy Mobile in production and can help
identify issues at scale.

Signed-off-by: JP Simard <jp@jpsim.com>