Disable global stats
This means that envoy bugs won't collect or emit stats, but these are
filtered out in our stats configuration anyway, so doesn't impact us
today.

In the future, if we do want to start emitting these, we'll need to
come up with a scheme to set up a global stats configuration that's not
tied to the engine's lifecycle, but we can defer on that for now.

Signed-off-by: JP Simard <jp@jpsim.com>