iOS: add SwiftPM example
And validate that it builds successfully on CI.

Ideally we'd have a whole suite of example integrations that validate
all our supported installation methods, but this should be a good start.

I reached out to the owner of https://catfacts.ninja to make sure this
usage is within the terms of service and will update to use something
else if necessary.

Signed-off-by: JP Simard <jp@jpsim.com>