swift: generate API docs as a DocC archive
And include in GitHub release artifacts

Updates rules_swift to pull in https://github.com/bazelbuild/rules_swift/pull/838

When users open the archive, they'll have rich docs they can explore
using Xcode's documentation navigator.

When we update to Xcode 13.4, we'll be able to generate HTML docs with
this approach and publish to the documentation website.

Signed-off-by: JP Simard <jp@jpsim.com>