Make CI jobs depend on Envoy build CI job (#2381)
Description: Platform specific CI jobs that run mobile applications have their timeouts set to 25 minutes. This is not enough to build Envoy Mobile itself and the apps that are used to verify that EM works as expected. Make CI jobs (that have timeouts equal to 25 minutes) depend on the 'main' CI job that's supposed to build Envoy Mobile binary and has a timeout of 90 minutes. This way, when platform specific CI jobs start they can reuse the cache that's populated by the CI job with timeout equal to 90 minutes.
Risk Level: N/A
Testing: N/A
Docs Changes: N/A
Release Notes: N/A

Signed-off-by: Rafal Augustyniak <raugustyniak@lyft.com>