key_value: Name the registered store in the PlatformKeyValueStoreConfig
The PlatformKeyValueStoreFactory used to take a PlatformInterfaceImpl
where the registered KV store was found in the API registry and used to
bridge the PlatformKeyValueStore with the native underlying store (e.g.
SharedPrefs on Android).

However, registered factory classes only have their default constructor
invoked, so the current PlatformKeyValueStoreFactory constructor taking
in the PlatformInterfaceImpl would not have worked.

Instead, we make the registered KV store name part of the
PlatformKeyValueStoreConfig and use the specified KV store name to
create the PlatformInstanceImpl inside the createStore function.  A
side-effect of this change is that the PlatformKeyValueStore owns the
PlatformInstanceImpl pointer.

Signed-off-by: Ali Beyad <abeyad@google.com>